+++ .eslintrc.js
@@ -5,7 +5,7 @@ module.exports = {
 	root: true,
 	parser: '@typescript-eslint/parser',
 	plugins: ['n8n-nodes-base'],
-	ignorePatterns: ['.eslintrc.js', 'dist/**', 'node_modules/**', 'gulpfile.js', 'scripts/**'],
+	ignorePatterns: ['.eslintrc.js', '.eslintrc.portal.js', 'dist/**', 'node_modules/**', 'gulpfile.js', 'scripts/**'],
 	extends: ['plugin:n8n-nodes-base/community'],
 	overrides: [
 		{